requirements.txt
Pillow
numpy
keras
tensorflow